Moving App.js to components, removing unneeded actions function, Removing unneeded Counter component, Fixing MovideDetails toggle
