--- src/components/counter/Counter.scss
@@ -1,11 +1,0 @@
-@import "../../styles/variables.scss";
-@import "../../styles/mixins.scss";
-
-.counter-input {
-  @include inputsStyles;
-  color: $white;
-  width: 5%;
-}
-.counter-button {
-  width: 5rem;
-}
