+++ src/components/movie-details/MovieDetails.scss
@@ -21,7 +21,7 @@
     .movieDetails-section-imageContainer {
       margin-right: 3.5rem;
       .movieDetails-section-image {
-        max-width: 18rem;
+        max-height: 25rem;
       }
     }
     .movieDetails-info {
