+++ src/components/shared/date-picker/DatePicker.scss
@@ -1,5 +1,5 @@
-@import "../../../../styles/variables.scss";
-@import "../../../../styles/mixins.scss";
+@import "../../../styles/variables.scss";
+@import "../../../styles/mixins.scss";
 
 .input-date-style {
   @include inputsStyles;
